pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sentinel*"]

[project]
name = "sentinel"
version = "2.0.0"
description = "Production monitoring, chaos engineering, and incident response system"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9",
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "httpx>=0.26",
    "black>=24.0",
    "isort>=5.13",
    "flake8>=7.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["sentinel"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 70
show_missing = true
